@@ -1,23 +1,23 @@
-name: regular_integration_tests
+name: google_sign_in_web_integration_tests
 publish_to: none
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  flutter: ">=1.27.0-0" # For integration_test from sdk
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  google_sign_in: ^4.5.3
+  http: ^0.13.0
+  js: ^0.6.3
   flutter_driver:
     sdk: flutter
   flutter_test:
     sdk: flutter
-  http: ^0.12.2
-  mockito: ^4.1.1
   integration_test:
-    path: ../../../integration_test
+    sdk: flutter
 
 dependency_overrides:
   google_sign_in_web:
