@@ -20,7 +20,7 @@ external set gapiOnloadCallback(Function callback);
 /// This is only exposed for testing. It shouldn't be accessed by users of the
 /// plugin as it could break at any point.
 @visibleForTesting
-const String kGapiOnloadCallbackFunctionName = "gapiOnloadCallback";
+const String kGapiOnloadCallbackFunctionName = 'gapiOnloadCallback';
 String _addOnloadToScript(String url) => url.startsWith('data:')
     ? url
     : '$url?onload=$kGapiOnloadCallbackFunctionName';
